Make Home/End keys behavior more standard
Currently, the Home / End keys navigate to the start / end of the entire prompt text, even if it's multiline. In the example text below, if the caret was at the very beginning of the text (before the first `a`), pressing End would navigate to the last `c`.

```
aaaa
bbbb
cccc
```

This is a bit unintuitive; most would expect End to navigate past the final `a` (and stay on the "first line" of the multiline text).

This PR fixes that, and also adds Control-Home and Control-End to navigate to the first/last characters of the multiline text.