Support returning IAsyncEnumerable from a keybinding
If a keybinding returns an IAsyncEnumerable, PrettyPrompt should consume
the IAsyncEnumerable and draw the results to the screen.

The use case here is an OpenAI integration I'm working on for
CSharpRepl. OpenAI streams the results in their API, which is
responsible for their "word by word" output UI. By accepting an
IAsyncEnumerable as a keybinding we can have a similar UI.

The contents that are streamed into our `Document` is done as a single
ChangeContext, so pressing Undo will undo the entire stream (which is
desirable in my opinion).