Fix altgr keyboard handling
Previously, PR #252 was attempting to fix the scenario where users can
have a Control-Alt-Space keybinding; the keybinding would be fired but
the space key would be erroneously typed.

However, that PR broke AltGr handling. It assumed that filtering Ctrl
to detect keybindings would be OK, and Alt keys would be unaffected.
This is not correct, though, because AltGr is represented via Ctrl-Alt,
so it actually broke every AltGr based character, like curly-brace on
AZERTY keyboards.

This commit undoes the original fix for PR #252 and fixes it in a better
way. Instead of assuming Ctrl is a keybinding, we use the actual
configured keybindings in the filter, rather than assuming Ctrl
represent a keybinding.